+++ pssh/clients/base/parallel.py
@@ -360,6 +360,9 @@ class BaseParallelSSHClient(object):
         :rtype: ``None``"""
         if output is None:
             output = self.get_last_output()
+            if output is None:
+                logger.info("No last output to join on - ``run_command`` has never been run.")
+                return
         elif not isinstance(output, list):
             raise ValueError("Unexpected output object type")
         cmds = [self.pool.spawn(self._join, host_out, timeout=timeout,